Refactor RaftStore interface
RaftStore's methods are refactored and a new method is added to make
the contract more explicit and easier to implement.

In addition, in some call sites, the order of operations are changed to
avoid the risk of divergence between the memory state and persisted state
if the persistence operations fail with exceptions.

https://github.com/MicroRaft/MicroRaft/issues/25